More dramatic refactoring, moving connection handling logic into its own module
This might compile but it definitely doesn't work!
